bus/hp_dio/98544: convert to anonymous namespace (nw)
